Move all of app/javascript/shared/ to app/javascript/lib/ [DEV-203]
It's not clear to me that there is any distinction between these two
folders, and having code semi-randomly divided between them is sort of
annoying. I prefer the shorter and more standard `lib` directory name,
so this change moves everything from `app/javascript/shared/` to
`app/javascript/lib/`.
