+++ app/javascript/admin_entrypoints/active_admin.ts
@@ -1,4 +1,4 @@
-import '@/shared/global_jquery';
+import '@/lib/global_jquery';
 import 'jquery-ui';
 import 'jquery-ui/ui/widgets/mouse';
 import '@activeadmin/activeadmin';
