+++ app/javascript/emoji_picker/components/BoostsForm.vue
@@ -40,9 +40,9 @@
 import { remove } from 'lodash-es';
 
 import { boosts } from '@/emoji_picker/emoji_data';
+import { http } from '@/lib/http';
 import { vueToast } from '@/lib/vue_toasts';
 import { api_json_preferences_path } from '@/rails_assets/routes';
-import { http } from '@/shared/http';
 import type { EmojiDataWithBoostedName } from '@/types';
 
 function addBoost() {
