+++ app/javascript/entrypoints/check_ins.ts
@@ -1,7 +1,7 @@
 import { createPinia } from 'pinia';
 
 import CheckIns from '@/check_ins/App.vue';
-import { renderApp } from '@/shared/customized_vue';
+import { renderApp } from '@/lib/customized_vue';
 
 const app = renderApp(CheckIns, '#check_ins_app');
 const pinia = createPinia();
