+++ app/javascript/entrypoints/emoji_picker.ts
@@ -1,7 +1,7 @@
 import Toast from 'vue-toastification';
 
 import EmojiPicker from '@/emoji_picker/EmojiPicker.vue';
-import { renderApp } from '@/shared/customized_vue';
+import { renderApp } from '@/lib/customized_vue';
 
 const app = renderApp(EmojiPicker);
