+++ app/javascript/entrypoints/groceries_app.ts
@@ -2,7 +2,7 @@ import { createPinia } from 'pinia';
 import Toast from 'vue-toastification';
 
 import Groceries from '@/groceries/Groceries.vue';
-import { renderApp } from '@/shared/customized_vue';
+import { renderApp } from '@/lib/customized_vue';
 
 const app = renderApp(Groceries);
