+++ app/javascript/entrypoints/home_app.ts
@@ -1,7 +1,7 @@
 import { createPinia } from 'pinia';
 
 import Home from '@/home/Home.vue';
-import { renderApp } from '@/shared/customized_vue';
+import { renderApp } from '@/lib/customized_vue';
 
 const app = renderApp(Home);
 const pinia = createPinia();
