+++ app/javascript/entrypoints/vue_playground.ts
@@ -1,4 +1,4 @@
+import { renderApp } from '@/lib/customized_vue';
 import Playground from '@/playground/Playground.vue';
-import { renderApp } from '@/shared/customized_vue';
 
 renderApp(Playground);
