+++ app/javascript/entrypoints/workout_app.ts
@@ -1,6 +1,6 @@
 import { createPinia } from 'pinia';
 
-import { renderApp } from '@/shared/customized_vue';
+import { renderApp } from '@/lib/customized_vue';
 import WorkoutApp from '@/workout/Workout.vue';
 
 const app = renderApp(WorkoutApp);
