+++ app/javascript/groceries/components/CheckInModal.vue
@@ -55,8 +55,8 @@ import { TYPE } from 'vue-toastification';
 
 import Modal from '@/components/Modal.vue';
 import { useGroceriesStore } from '@/groceries/store';
+import { useModalStore } from '@/lib/modal/store';
 import { vueToast } from '@/lib/vue_toasts';
-import { useModalStore } from '@/shared/modal/store';
 import type { Store } from '@/types';
 
 import CheckInItemsList from './CheckInItemsList.vue';
