+++ app/javascript/lib/customized_vue.ts
