+++ app/javascript/lib/global_jquery.ts
