+++ app/javascript/lib/helpers.ts
