+++ app/javascript/lib/index.d.ts
