+++ app/javascript/lib/pinia_custom_properties.d.ts
