+++ app/javascript/workout/components/ConfirmWorkoutModal.vue
@@ -35,8 +35,8 @@ import { ElButton, ElCheckbox } from 'element-plus';
 import { ref } from 'vue';
 
 import Modal from '@/components/Modal.vue';
+import { useModalStore } from '@/lib/modal/store';
 import { toast } from '@/lib/toasts';
-import { useModalStore } from '@/shared/modal/store';
 import { useWorkoutsStore } from '@/workout/store';
 
 const props = defineProps({
